#323 put router utils into own module
